Time: 55 ms (89.92%) | Memory: 11 MB (78.54%) - LeetSync

+++ 1890-sum-of-beauty-of-all-substrings/sum-of-beauty-of-all-substrings.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int beautySum(string s) {
         
-        int ans = 0;
+        long int ans = 0;
         
         for (int i=0; i< s.size();i++){
